% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_google_merch.R
\docType{data}
\name{daily_google_merch}
\alias{daily_google_merch}
\title{Daily Google Merchant Store Data}
\format{A data frame with 90 rows and 3 variables:
\describe{
  \item{date}{The dates (YYYYMMDD) of the active date range.}
  \item{transactions}{Transactions is the total number of completed purchases on your site.}
  \item{sessions}{Total number of Sessions within the date range. }
}}
\usage{
daily_google_merch
}
\description{
A dataset containing 90 days of the Google Merchant Store Demo data.
\url{https://support.google.com/analytics/answer/6367342?hl=en}
}
\keyword{datasets}
