% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_sourcemedium_wide.R
\docType{data}
\name{ga_sourcemedium_wide}
\alias{ga_sourcemedium_wide}
\title{Google Analytics SourceMedium Wide}
\format{A data frame with 2111 rows and 5 variables:
\describe{
  \item{ds}{datetime of the session}
  \item{...}{ga sourceMedium dimension}
}}
\usage{
ga_sourcemedium_wide
}
\description{
A dataset containing google analytics sessions with the extra dimension
of sourcemedium. Using an extra dimension (in addition to time based dims)
provides a way to subset data by a desired dimension. The dataset has been
spread into wide format.
}
\keyword{datasets}
