% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_analytics_sourcemedium.R
\docType{data}
\name{google_analytics_sourcemedium}
\alias{google_analytics_sourcemedium}
\title{Google Analytics SourceMedium Data}
\format{A data frame with 3920 rows and 4 variables:
\describe{
  \item{date}{date of the ga session}
  \item{hour}{hour of the ga session}
  \item{sourceMedium}{ga sourceMedium dimension}
  \item{sessions}{number of sessions metric from ga}
}}
\usage{
google_analytics_sourcemedium
}
\description{
A dataset containing google analytics sessions with the extra dimension
of sourcemedium. Using an extra dimension (in addition to time based dims)
provides a way to subset data by a desired dimension.
}
\keyword{datasets}
