% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_google_merch.R
\docType{data}
\name{hourly_google_merch}
\alias{hourly_google_merch}
\title{Hourly Google Merchant Store Data}
\format{A data frame with 2159 rows and 4 variables:
\describe{
  \item{date}{The dates (YYYYMMDD) of the active date range.}
  \item{hour}{The hour of the day. Values range from 00 to 23.}
  \item{transactions}{Transactions is the total number of completed purchases on your site.}
  \item{sessions}{Total number of Sessions within the date range. }
}}
\usage{
hourly_google_merch
}
\description{
A dataset containing 90 days of hourly Google Merchant Store Demo data.
\url{https://support.google.com/analytics/answer/6367342?hl=en}
}
\keyword{datasets}
