% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_pca.R
\name{sat_pca}
\alias{sat_pca}
\title{Satellite Interpretable PCA}
\usage{
sat_pca(x, min_nonzero = 1, min_val = 0.01)
}
\arguments{
\item{x}{matrix. A data frame. First column are date values,
then the numeric vectors for PCA}

\item{min_nonzero}{numeric. The minimum number of non-zero records for
exclusion.}

\item{min_val}{numeric. The minimum value in a column for exclusion.}
}
\value{
a list containing inputs and improtance measures.
}
\description{
Performs PCA on a matrix of observations and calculates a measure of variable
importance.
}
\details{
For a given matrix of values, the function will perform principal component
analysis and calculate a variable importance measure based on the PCA
scores and loadings to determine a local measure of variable importance.
}
\examples{
\dontrun{
data("google_analytics_sourcemedium")

library(tidyr)
library(dplyr)
library(lubridate)

ga_sourcemedium_wide <- google_analytics_sourcemedium \%>\%
mutate(ds = ymd_h(paste(date, hour))) \%>\%
spread(sourceMedium, sessions, fill = 0) \%>\%
select(-date, -hour)

x <- sat_pca(x = ga_sourcemedium_wide, min_nonzero = 1, min_val = 0.01)
 }
}
